refactor: extract shared rate limiter
